FInish the signup signin feature
Now add logic in your Index view to display the author’s name, but only if a user is signed in.
Sign in and create a few secret posts.
Test it out – sign out and go to the index page.
You should see a list of the posts but no author names.
Sign in and the author names should appear